requirements.txt
# requirements.txt
streamlit
pandas
mysql-connector-python
requests
groq
networkx
pyvis
matplotlib
plotly
numpy